pyproject.toml
[tool.poetry]
name = "notification-service"
version = "0.1.0"
description = ""
authors = ["katsiaryna-lysenka <katya.lysenko.1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pymongo = {extras = ["srv"], version = "^4.6.3"}
python-dotenv = "^1.0.1"
fastapi = {extras = ["all"], version = "^0.110.2"}
boto3 = "^1.34.89"
pika = "^1.3.2"
docker = "^7.0.0"
aiohttp = "^3.9.5"
api = "^0.0.7"
aio-pika = "^9.4.1"
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
